BackEnd/ClickKYCApi/requirements.txt
pycryptodome
requests
fastapi
sqlalchemy
pymysql
uvicorn
debugpy
pydantic[email]
httpx
cryptography
python-multipart
kafka-python
aiokafka
python-dotenv
asyncmy
uvicorn[standard]
jinja2 
weasyprint
slowapi
PyJWT